requirements.txt
streamlit==1.35.0
pandas==2.2.2
openai==1.34.0
python-dotenv==1.0.1
Pillow==10.3.0 